Refactor to combine client and proxy
I eliminated the proxy component and made it part of the same process as the
client, removing some complicated moving pieces and making the actual test
infrastructure more straightforward.

As part of this, the proxy and connections made through it are also now their
own types with methods and such, which makes everything a lot cleaner.

Timing data is still currently really wildly varying, which is troublesome.
Next steps are to reimplement testing sending a string N times and seeing if my
results stabilize.